chore: rewrap pragma comments + minor lint nits to satisfy ruff E501/RUF/N802
Pragma annotations added in the recent refactor wave exceeded the 120-char
line limit (E501); rewrap them as block comments above the guarded line.
Also: rename test_file_under_Boards_directory_detected_as_board to lowercase
(N802), prefix unused unpack tuples with _ (RUF059), drop redundant int(round())
in recall_steps fake (RUF046), and replace one ambiguous EN DASH in a docstring
(RUF002). No behaviour changes.

+++ tests/bdd/steps/recall_steps.py
@@ -130,7 +130,7 @@ def configured_checker(score: float) -> None:
         ) -> dict[str, object]:
             return {
                 "score": captured_score,
-                "passed": int(round(captured_score * 5)),
+                "passed": round(captured_score * 5),
                 "total": 5,
                 "timestamp": 0,
                 "detail": [],
